+++ cmd/falco/help.go
@@ -46,7 +46,6 @@ Usage:
 
 Subcommands:
     lint      : Run lint (default)
-    terraform : Run lint from terraform planned JSON
     stats     : Analyze VCL statistics
     simulate  : Run simulator server with provided VCLs
     test      : Run local testing for provided VCLs
@@ -103,6 +102,7 @@ Flags:
     -I, --include_path : Add include path
     -h, --help         : Show this help
     -r, --remote       : Connect with Fastly API
+    --proxy            : Enable actual proxy behavior
     -request           : Simulate request config
     -debug             : Enable debug mode
     --max_backends     : Override max backends limitation