+++ interpreter/statement.go
@@ -4,6 +4,7 @@ import (
 	"io"
 	"strings"
 
+	"github.com/k0kubun/pp"
 	"github.com/pkg/errors"
 	"github.com/ysugimoto/falco/ast"
 	"github.com/ysugimoto/falco/interpreter/assign"
@@ -33,6 +34,17 @@ func (i *Interpreter) ProcessBlockStatement(
 		// Call debugger
 		if debugState != DebugStepOut {
 			debugState = i.Debugger.Run(stmt)
+		}
+
+		// Find process marker and add flow if found
+		if stmt.GetMeta() == nil {
+			pp.Println(stmt)
+		}
+		if name, found := findProcessMark(stmt.GetMeta().Leading); found {
+			i.process.Flows = append(
+				i.process.Flows,
+				process.NewFlow(i.ctx, process.WithName(name), process.WithToken(stmt.GetMeta().Token)),
+			)
 		}
 
 		switch t := stmt.(type) {
@@ -282,12 +294,22 @@ func (i *Interpreter) ProcessCallStatement(stmt *ast.CallStatement, ds DebugStat
 	var state State
 	var err error
 	name := stmt.Subroutine.Value
+
 	if sub, ok := i.ctx.SubroutineFunctions[name]; ok {
+		// If mocked functional subroutine exists, use it
+		if mocked, ok := i.ctx.MockedFunctioncalSubroutines[name]; ok {
+			sub = mocked
+		}
+
 		_, state, err = i.ProcessFunctionSubroutine(sub, ds)
 		if err != nil {
 			return NONE, errors.WithStack(err)
 		}
 	} else if sub, ok = i.ctx.Subroutines[name]; ok {
+		// If mocked subroutine exists, use it
+		if mocked, ok := i.ctx.MockedSubroutines[name]; ok {
+			sub = mocked
+		}
 		state, err = i.ProcessSubroutine(sub, ds)
 		if err != nil {
 			return NONE, errors.WithStack(err)